+++ src/onegov.file/onegov/file/__init__.py
@@ -2,7 +2,6 @@ from onegov.file.collection import FileCollection, FileSetCollection
 from onegov.file.integration import DepotApp
 from onegov.file.models import (
     File,
-    FileMessage,
     FileSet,
     AssociatedFiles,
     SearchableFile
@@ -13,7 +12,6 @@ __all__ = (
     'DepotApp',
     'File',
     'FileCollection',
-    'FileMessage',
     'FileSet',
     'FileSetCollection',
     'SearchableFile',